backend/requirements.txt
fastapi==0.100.0
uvicorn==0.23.0
requests==2.31.0
pydantic==2.4.2
numpy==1.24.0
geopy==2.4.0
python-dotenv==1.0.0
scikit-learn==1.3.0
pandas==2.0.3
openmeteo-requests==1.1.0
requests-cache==1.1.0
retry-requests==2.0.0
joblib==1.3.2